Enum druid::Event

pub enum Event {
    Size(Size),
    MouseDown(MouseEvent),
    MouseUp(MouseEvent),
    MouseMoved(MouseEvent),
    HotChanged(bool),
    FocusChanged(bool),
}

An event, propagated downwards during event flow.

Events are things that happen that can change the state of widgets. An important category is events plumbed from the platform windowing system, which includes mouse and keyboard events, but also (in the future) status changes such as window focus changes.

Events can also be higher level concepts indicating state changes within the widget hierarchy, for example when a widget gains or loses focus or "hot" (also known as hover) status.

Events are a key part of what is called "event flow", which is basically the propagation of an event through the widget hierarchy through the event widget method. A container widget will generally pass the event to its children, mediated through the WidgetPod container, which is where most of the event flow logic is applied (especially the decision whether or not to propagate).

This enum is expected to grow considerably, as there are many, many different kinds of events that are relevant in a GUI.

Variants

Size(Size)

Called when the system has a file the application should open.

Most commonly this is in response to a request to show the file picker. Called on the root widget when the window size changes.

Discussion: it's not obvious this should be propagated to user widgets. It is propagated through the RootWidget and handled in the WindowPod, but after that it might be considered better to just handle it in layout.

The propagation logic of "just the root" requires a little bit of complexity and state in EventCtx, so if it's not useful it should be removed.

MouseDown(MouseEvent)

Called when a mouse button is pressed.

MouseUp(MouseEvent)

Called when a mouse button is released.

MouseMoved(MouseEvent)

Called when the mouse is moved.

The MouseMoved event is propagated to the active widget, if there is one, otherwise to hot widgets (see HotChanged).

The MouseMoved event is also the primary mechanism for widgets to set a cursor, for example to an I-bar inside a text widget. A simple tactic is for the widget to unconditionally call set_cursor in the MouseMoved handler, as MouseMove is only propagated to active or hot widgets.

HotChanged(bool)

Called when a key is pressed.

Note: the intent is for each physical key press to correspond to a single KeyDown event. This is sometimes different than the raw events provided by the platform. In particular, Windows sends one or both of WM_KEYDOWN (a raw key code) and WM_CHAR (the Unicode value), depending on the actual key. Called when a key is released.

Because of repeat, there may be a number KeyDown events before a corresponding KeyUp is sent. Called when a paste command is received. Called when the mouse wheel or trackpad is scrolled. Called when the "hot" status changes.

See is_hot for discussion about the hot status.

FocusChanged(bool)

Called when the focus status changes.

See has_focus for discussion about the focus status.

Methods

impl Event

pub fn transform_scroll(&self, offset: Vec2, viewport: Rect) -> Option<Event>

Transform the event for the contents of a scrolling container.

pub(crate) fn recurse(&self) -> bool

Whether the event should be propagated from parent to children.