Module druid::env

An environment which is passed downward into the widget tree.

Structs

Env
An environment passed down through all widget traversals.

Key
A typed key.

Enums

Value
A dynamic type representing all values that can be stored in an environment.

Constants

BORDER

FONT_HEIGHT

FONT_NAME

LABEL_COLOR

TEXT_SIZE_NORMAL

WINDOW_BACKGROUND_COLOR

WINDOW_HEIGHT

WINDOW_WIDTH

Traits

ValueType
Values which can be stored in an environment.

Type Definitions

EnvError
The error type for environment access.

MaxStringValue

String