Struct druid::env::Env

pub struct Env {}

An environment passed down through all widget traversals.

All widget methods have access to an environment, and it is passed downwards during traversals.

A widget can retrieve theme parameters (colors, dimensions, etc.). In addition, it can pass custom data down to all descendants. An important example of the latter is setting a value for enabled/disabled status so that an entire subtree can be disabled ("grayed out") with one setting. #[derive(Clone)]

Methods

impl Env

pub fn get<'a, V: ValueType<'a>>(&'a self, _key: Key<V>) -> V

Gets a value from the environment, expecting it to be present.

Note that the return value is a reference for "expensive" types such as strings, but an ordinary value for "cheap" types such as numbers and colors.

Panics

Panics if the key is not found, or if it is present with the wrong type.

pub fn try_get<'a, V: ValueType<'a>>(&'a self, _key: Key<V>) -> Option<V>

Gets a value from the environment.

Panics

Panics if the value for the key is found, but has the wrong type.

pub fn adding<'a, V: ValueType<'a>>(
self,
_key: Key<V>,
_value: impl Into<V::Owned>
) -> Env

Adds a key/value, acting like a builder.

pub fn set<'a, V: ValueType<'a>>(
&'a mut self,
_key: Key<V>,
_value: impl Into<V::Owned>
)

Sets a value in an environment.

Panics

Panics if the environment already has a value for the key, but it is of a different type.