Module druid::localization

Localization handling.

Localization is backed by Fluent, via fluent-rs.

In Druid, the main way you will deal with localization is via the LocalizedString struct.

You construct a LocalizedString with a key, which identifies a 'message' in your .flt files. If your string requires arguments, you supply it with closures that can extract those arguments from the current Env and Data.

At runtime, you resolve your LocalizedString into an actual string, passing it the current Env and Data.

Structs

ArgSource
Wraps a closure that generates an argument for localization.

LocalizedString
A string that can be localized based on the current locale.

Type Definitions

ArgClosure

MaxLocalizedArg

MaxLocalizedString

String

Vec