Struct druid::EventCtx

pub struct EventCtx<'a, D: Data + 'static + Default> {
    win_ctx: &'a mut dyn WinCtx,
    window_id: WindowId,
    window: &'a WindowHandle<DruidHandler<D>>,
    base_state: &'a mut BaseState,
    had_active: bool,
    is_handled: bool,
    is_root: bool,
}

A mutable context provided to event handling methods of widgets.

Widgets should call invalidate whenever an event causes a change in the widget's appearance, to schedule a repaint.

Fields
win_ctx: &'a mut dyn WinCtxwindow_id: WindowId

Commands submitted to be run after this event.

window: &'a WindowHandle<DruidHandler<D>>base_state: &'a mut BaseStatehad_active: boolis_handled: boolis_root: bool
Methods

impl<'a, D: Data + 'static + Default> EventCtx<'a, D>

pub fn invalidate(&mut self)

Invalidate.

Right now, it just invalidates the entire window, but we'll want finer grained invalidation before long.

pub fn text(&mut self) -> &mut Text

Get an object which can create text layouts.

pub fn set_active(&mut self, active: bool)

Set the "active" state of the widget.

See BaseState::is_active.

pub fn is_hot(&self) -> bool

Query the "hot" state of the widget.

See BaseState::is_hot.

pub fn is_active(&self) -> bool

Query the "active" state of the widget.

This is the same state set by set_active and is provided as a convenience.

pub fn window(&self) -> &WindowHandle<DruidHandler<D>>

Returns a reference to the current WindowHandle.

Note: we're in the process of migrating towards providing functionality provided by the window handle in mutable contexts instead. If you're considering a new use of this method, try adding it to WinCtx and plumbing it through instead.

pub fn set_handled(&mut self)

Set the event as "handled", which stops its propagation to other widgets.

pub fn is_handled(&self) -> bool

Determine whether the event has been handled by some other widget.

pub fn has_focus(&self) -> bool

Query the focus state of the widget.

See BaseState::has_focus.

pub fn request_focus(&mut self)

Request keyboard focus.

Discussion question: is method needed in contexts other than event?

pub fn size(&self) -> Size

Returns the layout size of the current widget.

pub fn window_id(&self) -> WindowId

Get the window id.