Struct druid::WidgetPod

pub struct WidgetPod<T: Data + 'static + Default, W: Widget<T>> {
    state: BaseState,
    old_data: Option<T>,
    env: Option<Env>,
    inner: W,
}

A container for one widget in the hierarchy.

Generally, container widgets don't contain other widgets directly, but rather contain a WidgetPod, which has additional state needed for layout and for the widget to participate in event flow.

This struct also contains the previous data for a widget, which is essential for the update method, both to decide when the update needs to propagate, and to provide the previous data so that a widget can process a diff between the old value and the new.

Fields
state: BaseStateold_data: Option<T>env: Option<Env>inner: W
Methods

impl<T: Data + 'static + Default, W: Widget<T>> WidgetPod<T, W>

pub fn new(inner: W) -> WidgetPod<T, W>

Create a new widget pod.

In a widget hierarchy, each widget is wrapped in a WidgetPod so it can participate in layout and event flow. The process of adding a child widget to a container should call this method.

pub fn widget(&self) -> &W

Return a reference to the inner widget.

pub fn widget_mut(&mut self) -> &mut W

Return a mutable reference to the inner widget.

pub fn set_layout_rect(&mut self, layout_rect: Rect)

Set layout rectangle.

Intended to be called on child widget in container's layout implementation.

pub fn get_layout_rect(&self) -> Rect

Get the layout rectangle.

This will be same value as set by set_layout_rect.

pub fn paint(&mut self, paint_ctx: &mut PaintCtx, data: &T, env: &Env)

Paint a child widget.

Generally called by container widgets as part of their paint method.

Note that this method does not apply the offset of the layout rect. If that is desired, use paint_with_offset instead.

pub fn paint_with_offset(
&mut self,
paint_ctx: &mut PaintCtx,
data: &T,
env: &Env
)

Paint the widget, translating it by the origin of its layout rectangle.

This will recursively paint widgets, stopping if a widget's layout rect is outside of the currently visible region.

pub fn paint_with_offset_always(
&mut self,
paint_ctx: &mut PaintCtx,
data: &T,
env: &Env
)

Paint the widget, even if its layout rect is outside of the currently visible region.

fn paint_with_offset_impl(
&mut self,
paint_ctx: &mut PaintCtx,
data: &T,
env: &Env,
paint_if_not_visible: bool
)

Shared implementation that can skip drawing non-visible content.

pub fn layout(
&mut self,
layout_ctx: &mut LayoutCtx,
bc: &BoxConstraints,
data: &T,
env: &Env
) -> Size

Compute layout of a widget.

Generally called by container widgets as part of their layout method.

pub fn event(
&mut self,
ctx: &mut EventCtx<T>,
event: &Event,
data: &mut T,
env: &Env
)

Propagate an event.

Generally the event method of a container widget will call this method on all its children. Here is where a great deal of the event flow logic resides, particularly whether to continue propagating the event.

pub fn update(&mut self, ctx: &mut UpdateCtx<T>, data: &T, env: &Env)

Propagate a data update.

Generally called by container widgets as part of their update method.