Struct druid::WindowDesc

pub struct WindowDesc<T: Data + 'static + Default, W: Widget<T> + 'static> {
    pub(crate) root_builder: fn() -> W,
    pub(crate) size: Option<Size>,
    pub id: WindowId,
    phantom_data: PhantomData<T>,
}

A function that can create a widget. A description of a window to be instantiated.

This includes a function that can build the root widget, as well as other window properties such as the title.

Fields
root_builder: fn() -> Wsize: Option<Size>id: WindowId

The WindowId that will be assigned to this window.

This can be used to track a window from when it is launched and when it actually connects.

phantom_data: PhantomData<T>
Methods

impl<T: Data + 'static + Default, W: Widget<T> + 'static> WindowDesc<T, W>

pub fn new(root: fn() -> W) -> WindowDesc<T, W>

Create a new WindowDesc, taking a funciton that will generate the root Widget for this window.

It is possible that a WindowDesc can be reused to launch multiple windows.

pub fn window_size(self, size: impl Into<Size>) -> Self

Set the window size at creation

You can pass in a tuple (width, height) or kurbo::Size e.g. to create a window 1000px wide and 500px high

window.window_size((1000.0, 500.0));

pub(crate) fn build_native(
&self,
state: &mut AppState<T>
) -> Result<WindowHandle<DruidHandler<T>>, PlatformError>

Attempt to create a platform window from this WindowDesc.