Module druid::widget

Common widgets.

Modules

align
A widget that aligns its child (for example, centering it).

button
A button widget.

flex
A widget that arranges its children in a one-dimensional array.

label
A label widget.

padding
A widget that just adds padding during layout.

widgetbox
WidgetBox contains a Widget. Allows for dynamic dispatch with static Widgets in [no_std].

Structs

Align
A widget that aligns its child.

Button
A button with a text label.

Column
A builder for a column widget that can contain flex children.

Flex
A container with either horizontal or vertical layout.

Label
A label that displays some text.

Padding
A widget that just adds padding around its child.

Row
A builder for a row widget that can contain flex children.

WidgetBox
Boxed version of a Widget

Enums

LabelText
The text for the label; either a localized or a specific string.

WidgetType
Enum to store each Widget

Constants

MAX_WIDGETS
Max number of Widgets on embedded platforms

Traits

GlobalWidgets
Specialised Trait for handling static Widgets on embedded platforms

Functions

get_widget_id
Assign a unique Widget ID

Type Definitions

WidgetId
Widgets are identified by an 8-bit ID