Module druid::widget::flex

A widget that arranges its children in a one-dimensional array.

Structs

ChildWidget

Column
A builder for a column widget that can contain flex children.

Flex
A container with either horizontal or vertical layout.

Params

Row
A builder for a row widget that can contain flex children.

Enums

Axis

Type Definitions

MaxWidgets

Vec