Struct druid::widget::Flex

pub struct Flex<T: Data + 'static + Default> {
    id: WidgetId,
    direction: Axis,
    children: Vec<ChildWidget<T>, U8>,
}

A container with either horizontal or vertical layout.

Fields
id: WidgetIddirection: Axischildren: Vec<ChildWidget<T>, U8>
Methods

impl<T: Data + 'static + Default> Flex<T>

pub fn add_child<W: Widget<T> + Clone>(&mut self, child: W, flex: f64)

Add a child widget.

If flex is zero, then the child is non-flex. It is given the same constraints on the "minor axis" as its parent, but unconstrained on the "major axis".

If flex is non-zero, then all the space left over after layout of the non-flex children is divided up, in proportion to the flex value, among the flex children.