Module druid::widget::widgetbox

WidgetBox contains a Widget. Allows for dynamic dispatch with static Widgets in [no_std].

Structs

WidgetBox
Boxed version of a Widget

Enums

WidgetType
Enum to store each Widget

Constants

MAX_WIDGETS
Max number of Widgets on embedded platforms

Statics

WIDGET_ID
Unique Widget ID

Traits

GlobalWidgets
Specialised Trait for handling static Widgets on embedded platforms

Functions

get_widget_id
Assign a unique Widget ID

Type Definitions

WidgetId
Widgets are identified by an 8-bit ID