Module druid::win_handler

The implementation of the WinHandler trait (druid-shell integration).

Structs

AppState
State shared by all windows in the UI.

DruidHandler
The struct implements the druid-shell WinHandler trait.

SingleWindowState
Everything required for a window to handle an event.

WindowState
Per-window state not owned by user code.

Windows
All active windows.

Constants

MAX_WINDOWS
Max number of Windows supported. i=0 is not used, so MAX_WINDOWS should be 1 more than max number of Windows.

Traits

GlobalWindows
Specialised Trait for handling static Windows, Window Handlers and Application Data on embedded platforms