Struct druid::win_handler::DruidHandler

pub struct DruidHandler<T: Data + 'static + Default> {
    pub window_id: WindowId,
    pub phantom: PhantomData<T>,
}

The struct implements the druid-shell WinHandler trait.

One DruidHandler exists per window.

This is something of an internal detail and possibly we don't want to surface it publicly.

Fields
window_id: WindowId

The shared app state. The id for the current window.

phantom: PhantomData<T>
Methods

impl<T: Data + 'static + Default> DruidHandler<T>

pub(crate) fn new_shared(window_id: WindowId) -> DruidHandler<T>

Note: the root widget doesn't go in here, because it gets added to the app state.

fn do_event(&mut self, event: Event, win_ctx: &mut dyn WinCtx) -> bool

Send an event to the widget hierarchy.

Returns true if the event produced an action.

This is principally because in certain cases (such as keydown on Windows) the OS needs to know if an event was handled.